+++ src/LibForumData/private/MemoryRepositoryUser.cpp
@@ -13,17 +13,38 @@ using namespace Forum::Configuration;
 using namespace Forum::Entities;
 using namespace Forum::Helpers;
 using namespace Forum::Repository;
+using namespace Forum::Authorization;
 
-MemoryRepositoryUser::MemoryRepositoryUser(MemoryStoreRef store) : MemoryRepositoryBase(std::move(store)),
-    validUserNameRegex(boost::make_u32regex("^[[:alnum:]]+[ _-]*[[:alnum:]]+$"))
-{}
+MemoryRepositoryUser::MemoryRepositoryUser(MemoryStoreRef store, UserAuthorizationRef authorization) 
+    : MemoryRepositoryBase(std::move(store)), validUserNameRegex(boost::make_u32regex("^[[:alnum:]]+[ _-]*[[:alnum:]]+$")),
+    authorization_(std::move(authorization))
+{
+    if ( ! authorization_)
+    {
+        throw std::runtime_error("Authorization implementation not provided");
+    }
+}
 
 StatusCode MemoryRepositoryUser::getUsers(OutStream& output, RetrieveUsersBy by) const
 {
+    StatusWriter status(output, StatusCode::OK);
+
     PerformedByWithLastSeenUpdateGuard performedBy;
+    AuthorizationStatus authorizationStatus;
 
     collection().read([&](const EntityCollection& collection)
     {
+        auto& currentUser = performedBy.get(collection, store());
+
+        authorizationStatus = authorization_->getUsers(currentUser);
+        if (AuthorizationStatusCode::OK != authorizationStatus.code)
+        {
+            status = authorizationStatus.code;
+            return;
+        }
+
+        status.disable();
+
         auto pageSize = getGlobalConfig()->user.maxUsersPerPage;
         auto& displayContext = Context::getDisplayContext();
 
@@ -52,9 +73,9 @@ StatusCode MemoryRepositoryUser::getUsers(OutStream& output, RetrieveUsersBy by)
         }
 
         if ( ! Context::skipObservers())
-            readEvents().onGetUsers(createObserverContext(performedBy.get(collection, store())));
+            readEvents().onGetUsers(createObserverContext(currentUser));
     });
-    return StatusCode::OK;
+    return status;
 }
 
 StatusCode MemoryRepositoryUser::getUserById(const IdType& id, OutStream& output) const
@@ -62,9 +83,12 @@ StatusCode MemoryRepositoryUser::getUserById(const IdType& id, OutStream& output
     StatusWriter status(output, StatusCode::OK);
 
     PerformedByWithLastSeenUpdateGuard performedBy;
+    AuthorizationStatus authorizationStatus;
 
     collection().read([&](const EntityCollection& collection)
                       {
+                          auto& currentUser = performedBy.get(collection, store());
+
                           const auto& index = collection.usersById();
                           auto it = index.find(id);
                           if (it == index.end())
@@ -72,13 +96,19 @@ StatusCode MemoryRepositoryUser::getUserById(const IdType& id, OutStream& output
                               status = StatusCode::NOT_FOUND;
                               return;
                           }
-                          else
+
+                          authorizationStatus = authorization_->getUserById(currentUser, **it);
+                          if (AuthorizationStatusCode::OK != authorizationStatus.code)
                           {
-                              status.disable();
-                              writeSingleValueSafeName(output, "user", **it);
+                              status = authorizationStatus.code;
+                              return;
                           }
+
+                          status.disable();
+                          writeSingleValueSafeName(output, "user", **it);
+                          
                           if ( ! Context::skipObservers())
-                              readEvents().onGetUserById(createObserverContext(performedBy.get(collection, store())), id);
+                              readEvents().onGetUserById(createObserverContext(currentUser), id);
                       });
     return status;
 }
@@ -88,6 +118,7 @@ StatusCode MemoryRepositoryUser::getUserByName(const StringView& name, OutStream
     StatusWriter status(output, StatusCode::OK);
 
     PerformedByWithLastSeenUpdateGuard performedBy;
+    AuthorizationStatus authorizationStatus;
 
     //keep a string around in order to avoid dynamic memory allocation on each call
     //when converting from StringView to std::string in order to use index.find();
@@ -101,6 +132,8 @@ StatusCode MemoryRepositoryUser::getUserByName(const StringView& name, OutStream
     
     collection().read([&](const EntityCollection& collection)
                       {
+                          auto& currentUser = performedBy.get(collection, store());
+
                           const auto& index = collection.usersByName();
                           auto it = index.find(toString(name, nameString));
                           if (it == index.end())
@@ -108,14 +141,19 @@ StatusCode MemoryRepositoryUser::getUserByName(const StringView& name, OutStream
                               status = StatusCode::NOT_FOUND;
                               return;
                           }
-                          else
+
+                          authorizationStatus = authorization_->getUserByName(currentUser, **it);
+                          if (AuthorizationStatusCode::OK != authorizationStatus.code)
                           {
-                              status.disable();
-                              writeSingleValueSafeName(output, "user", **it);
+                              status = authorizationStatus.code;
+                              return;
                           }
+
+                          status.disable();
+                          writeSingleValueSafeName(output, "user", **it);
+
                           if ( ! Context::skipObservers())
-                              readEvents().onGetUserByName(createObserverContext(performedBy.get(collection, store())),
-                                                           name);
+                              readEvents().onGetUserByName(createObserverContext(currentUser), name);
                       });
     return status;
 }
@@ -133,15 +171,25 @@ StatusCode MemoryRepositoryUser::addNewUser(const StringView& name, OutStream& o
     }
 
     PerformedByWithLastSeenUpdateGuard performedBy;
+    AuthorizationStatus authorizationStatus;
 
     collection().write([&](EntityCollection& collection)
                        {
                            auto nameString = toString(name);
 
+                           auto currentUser = performedBy.getAndUpdate(collection);
+
                            auto& indexByName = collection.users().get<EntityCollection::UserCollectionByName>();
                            if (indexByName.find(nameString) != indexByName.end())
                            {
                                status = StatusCode::ALREADY_EXISTS;
+                               return;
+                           }
+
+                           authorizationStatus = authorization_->addNewUser(*currentUser, name);
+                           if (AuthorizationStatusCode::OK != authorizationStatus.code)
+                           {
+                               status = authorizationStatus.code;
                                return;
                            }
 
@@ -153,8 +201,7 @@ StatusCode MemoryRepositoryUser::addNewUser(const StringView& name, OutStream& o
                            collection.users().insert(user);
 
                            if ( ! Context::skipObservers())
-                               writeEvents().onAddNewUser(createObserverContext(*performedBy.getAndUpdate(collection)),
-                                                          *user);
+                               writeEvents().onAddNewUser(createObserverContext(*currentUser), *user);
 
                            status.writeNow([&](auto& writer)
                                            {
@@ -178,14 +225,24 @@ StatusCode MemoryRepositoryUser::changeUserName(const IdType& id, const StringVi
         return status = validationCode;
     }
     PerformedByWithLastSeenUpdateGuard performedBy;
+    AuthorizationStatus authorizationStatus;
 
     collection().write([&](EntityCollection& collection)
                        {
+                           auto currentUser = performedBy.getAndUpdate(collection);
+
                            auto& indexById = collection.users().get<EntityCollection::UserCollectionById>();
                            auto it = indexById.find(id);
                            if (it == indexById.end())
                            {
                                status = StatusCode::NOT_FOUND;
+                               return;
+                           }
+
+                           authorizationStatus = authorization_->changeUserName(*currentUser, **it, newName);
+                           if (AuthorizationStatusCode::OK != authorizationStatus.code)
+                           {
+                               status = authorizationStatus.code;
                                return;
                            }
 
@@ -202,7 +259,7 @@ StatusCode MemoryRepositoryUser::changeUserName(const IdType& id, const StringVi
                                                          user.name() = std::move(newNameString);
                                                      });
                            if ( ! Context::skipObservers())
-                               writeEvents().onChangeUser(createObserverContext(*performedBy.getAndUpdate(collection)),
+                               writeEvents().onChangeUser(createObserverContext(*currentUser),
                                                           **it, User::ChangeType::Name);
                        });
     return status;
@@ -222,9 +279,12 @@ StatusCode MemoryRepositoryUser::changeUserInfo(const IdType& id, const StringVi
     }
 
     PerformedByWithLastSeenUpdateGuard performedBy;
+    AuthorizationStatus authorizationStatus;
 
     collection().write([&](EntityCollection& collection)
                        {
+                           auto currentUser = performedBy.getAndUpdate(collection);
+
                            auto& indexById = collection.users().get<EntityCollection::UserCollectionById>();
                            auto it = indexById.find(id);
                            if (it == indexById.end())
@@ -232,13 +292,21 @@ StatusCode MemoryRepositoryUser::changeUserInfo(const IdType& id, const StringVi
                                status = StatusCode::NOT_FOUND;
                                return;
                            }
+
+                           authorizationStatus = authorization_->changeUserInfo(*currentUser, **it, newInfo);
+                           if (AuthorizationStatusCode::OK != authorizationStatus.code)
+                           {
+                               status = authorizationStatus.code;
+                               return;
+                           }
+
                            collection.modifyUser(it, [&newInfo](User& user)
                                                      {
                                                          user.info() = toString(newInfo);
                                                      });
 
                            if ( ! Context::skipObservers())
-                               writeEvents().onChangeUser(createObserverContext(*performedBy.getAndUpdate(collection)),
+                               writeEvents().onChangeUser(createObserverContext(*currentUser),
                                                           **it, User::ChangeType::Info);
                        });
     return status;
@@ -252,9 +320,12 @@ StatusCode MemoryRepositoryUser::deleteUser(const IdType& id, OutStream& output)
         return status = StatusCode::INVALID_PARAMETERS;
     }
     PerformedByWithLastSeenUpdateGuard performedBy;
+    AuthorizationStatus authorizationStatus;
 
     collection().write([&](EntityCollection& collection)
                        {
+                           auto currentUser = performedBy.getAndUpdate(collection);
+
                            auto& indexById = collection.users().get<EntityCollection::UserCollectionById>();
                            auto it = indexById.find(id);
                            if (it == indexById.end())
@@ -262,10 +333,17 @@ StatusCode MemoryRepositoryUser::deleteUser(const IdType& id, OutStream& output)
                                status = StatusCode::NOT_FOUND;
                                return;
                            }
+
+                           authorizationStatus = authorization_->deleteUser(*currentUser, **it);
+                           if (AuthorizationStatusCode::OK != authorizationStatus.code)
+                           {
+                               status = authorizationStatus.code;
+                               return;
+                           }
+
                            //make sure the user is not deleted before being passed to the observers
                            if ( ! Context::skipObservers())
-                               writeEvents().onDeleteUser(createObserverContext(*performedBy.getAndUpdate(collection)),
-                                                          **it);
+                               writeEvents().onDeleteUser(createObserverContext(*currentUser), **it);
 
                            collection.deleteUser(it);
                        });
